Move tests to `util` crate
